Enable network operations over neutron private nets
Enables communication between the savanna api server and the spawned
VMs over the specific network namespace associated with the selected
network.  The SSH and HTTP communication will be conducted over
netcat based subprocesses.  This feature is enabled when the
configuration indicates that neutron networking is enabled while
the use of public IPs is set to false.

Implements blueprint neutron-private-net-provisioning

Change-Id: I61d61a7ef6eccb04274549883dde489641d1da9b