Adding basic bandit config
This change adds a basic bandit config for sahara. It can be invoked by
running the tox environment for bandit; `$ tox -e bandit`. The tests are
based on the default bandit configuration with a few blacklist tests as
well. This is intended as a starting point for using bandit with sahara
and it should be revisited to improve the testing as more is learned
about the specific needs of the sahara code base.

* adding bandit.yaml configuration
* adding bandit to tox.ini
* adding a requirements file for bandit tox environment

Change-Id: I4b03f04dca80b146fdbae31a6b6011e78380d665