RHEL: fix distro detection and EPEL configuration
Distribution detection was working and it probably broke with RHEL 7.4.
This issue affects at least the MapR plugin, which installs few packages
at run-time and it fails to find the distribution.
A new field was added (internal_name) and used only for the comparison
because the 'name' proprty is used as parameter in other scripts and
changing it would touch a lot of code.

Also, fix the installation of EPEL repository, which was fixed
some time ago (from Pike) so that it did not depend on a specific
hardcoded version, but the fix worked only on CentOS, not on RHEL
where the epel-release package is not available.

Closes-Bug: #1740511
Related-Bug: #1657048
Change-Id: I21dc029f7bf556cd7dc0150ac2d7609494dba2fc