Fix the detection of scala version (now https)
The www.scala-lang.org switched to https and the curl call did not
account for this scenario, so it failed.
Switch the address to https and as extra measure add also the -L
argument, which forces curl to follow the redirects.

Switch also the download URL to https.

Story: 2001964
Task: 15088
Change-Id: I9d28c2ad56292998a6b7aaea421b98e136fbf6e1