Migrate away from oslo_i18n.enable_lazy()
The method is known to be problematic, and it is going to be
deprecated:
http://lists.openstack.org/pipermail/openstack-operators/2018-November/016139.html

Moreover, its usage in Sahara produces some nasty side-effects
when Python 3.x is used. The provisioning of the clusters ends
abruptly when sahara-engine tries to update the status in the
database, with a failure inside oslo_i18n.

Change-Id: I4ace7936251449e468fc2153bb048a2b90940539